Add dummy git command
If the course has commands like "git submodule update --init" in
old build.sh files, then the git command does not cause errors
(because git is not installed). Courses are encouraged to
remove the git commands from build.sh in A+ v1.12 and
the new Git Manager.

Git is not installed in the build containers since it is not
needed there.

Git Manager prepares git submodules before building the course
(running the build container).